separate out a wrapper SetGroups class and refactor SetGroupsAZFP and SetGroupsEK60, remove bins_to_avg and time_to_avg params from AZFP unpacking
