+++ echopype/convert/__init__.py
@@ -4,4 +4,3 @@ Include code to unpack manufacturer-specific data files into an interoperable ne
 from .convert import Convert
 from .ek60 import ConvertEK60
 from .azfp import ConvertAZFP
-from .set_groups import SetGroupsBase
