+++ echopype/model/azfp.py
@@ -92,8 +92,3 @@ class ModelAZFP(ModelBase):
             self.TS.to_dataset(name="TS").to_netcdf(path=self.TS_path, mode="w")
 
         ds_beam.close()
-
-    def get_MVBS(self):
-        with xr.open_dataset(self.file_path, group='Beam') as ds_beam:
-            super().get_MVBS('Sv', ds_beam.bins_to_avg, ds_beam.time_to_avg)
-        pass
